Uploader/requirements.txt
astroid==2.3.3
asttokens==2.0.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
colorama==0.4.3
docutils==0.15.2
ecdsa==0.15
esptool==2.8
ffmpeg==1.4
firebase-admin==4.5.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
httplib2==0.19.1
idna==2.10
isort==4.3.21
jedi==0.15.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
mutagen==1.45.1
mypy==0.761
mypy-extensions==0.4.3
packaging==20.9
panda3d==1.10.7
parso==0.5.2
Pillow==7.2.0
proto-plus==1.18.1
protobuf==3.15.8
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==1.9.6
pylint==2.4.4
pyparsing==2.4.7
pyserial==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
thonny==3.2.7
tkScrolledFrame==1.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.26.4
VPi==0.1.2
wrapt==1.11.2
